1. added quick and dirty first version of photo transmitting mechanism.

+++ PhotoSharing/src/ca/jvsh/photosharing/FileDataTransfer.java
@@ -6,10 +6,10 @@ public class FileDataTransfer implements Serializable
 {
 	private static final long	serialVersionUID	= 1L;
 	public boolean tcpOnly;
-	public int fileSize;
+	public long fileSize;
 	public String fileName;
 	
-	public FileDataTransfer (boolean tcpOnly, int fileSize, String fileName)
+	public FileDataTransfer (boolean tcpOnly, long fileSize, String fileName)
 	{
 		this.tcpOnly = tcpOnly;
 		this.fileSize=fileSize;
